+++ exporter/opentelemetry-exporter-zipkin/tests/test_zipkin_exporter.py
@@ -18,12 +18,21 @@ import unittest
 from unittest.mock import MagicMock, patch
 
 from opentelemetry import trace as trace_api
-from opentelemetry.exporter.zipkin import ZipkinSpanExporter
+from opentelemetry.configuration import Configuration
+from opentelemetry.exporter.zipkin import (
+    SPAN_KIND_MAP_JSON,
+    SPAN_KIND_MAP_PROTOBUF,
+    TRANSPORT_FORMAT_JSON,
+    TRANSPORT_FORMAT_PROTOBUF,
+    ZipkinSpanExporter,
+    nsec_to_usec_round,
+)
+from opentelemetry.exporter.zipkin.gen import zipkin_pb2
 from opentelemetry.sdk import trace
 from opentelemetry.sdk.trace import Resource
 from opentelemetry.sdk.trace.export import SpanExportResult
 from opentelemetry.sdk.util.instrumentation import InstrumentationInfo
-from opentelemetry.trace import TraceFlags
+from opentelemetry.trace import SpanKind, TraceFlags
 from opentelemetry.trace.status import Status, StatusCode
 
 
@@ -49,11 +58,17 @@ class TestZipkinSpanExporter(unittest.TestCase):
     def tearDown(self):
         if "OTEL_EXPORTER_ZIPKIN_ENDPOINT" in os.environ:
             del os.environ["OTEL_EXPORTER_ZIPKIN_ENDPOINT"]
+        if "OTEL_EXPORTER_ZIPKIN_TRANSPORT_FORMAT" in os.environ:
+            del os.environ["OTEL_EXPORTER_ZIPKIN_TRANSPORT_FORMAT"]
+        Configuration()._reset()  # pylint: disable=protected-access
 
     def test_constructor_env_var(self):
         """Test the default values assigned by constructor."""
         url = "https://foo:9911/path"
         os.environ["OTEL_EXPORTER_ZIPKIN_ENDPOINT"] = url
+        os.environ[
+            "OTEL_EXPORTER_ZIPKIN_TRANSPORT_FORMAT"
+        ] = TRANSPORT_FORMAT_PROTOBUF
         service_name = "my-service-name"
         port = 9911
         exporter = ZipkinSpanExporter(service_name)
@@ -65,6 +80,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
         self.assertEqual(exporter.ipv6, ipv6)
         self.assertEqual(exporter.url, url)
         self.assertEqual(exporter.port, port)
+        self.assertEqual(exporter.transport_format, TRANSPORT_FORMAT_PROTOBUF)
 
     def test_constructor_default(self):
         """Test the default values assigned by constructor."""
@@ -74,12 +90,14 @@ class TestZipkinSpanExporter(unittest.TestCase):
         ipv4 = None
         ipv6 = None
         url = "http://localhost:9411/api/v2/spans"
+        transport_format = TRANSPORT_FORMAT_JSON
 
         self.assertEqual(exporter.service_name, service_name)
         self.assertEqual(exporter.port, port)
         self.assertEqual(exporter.ipv4, ipv4)
         self.assertEqual(exporter.ipv6, ipv6)
         self.assertEqual(exporter.url, url)
+        self.assertEqual(exporter.transport_format, transport_format)
 
     def test_constructor_explicit(self):
         """Test the constructor passing all the options."""
@@ -88,8 +106,14 @@ class TestZipkinSpanExporter(unittest.TestCase):
         ipv4 = "1.2.3.4"
         ipv6 = "2001:0db8:85a3:0000:0000:8a2e:0370:7334"
         url = "https://opentelemetry.io:15875/myapi/traces?format=zipkin"
+        transport_format = TRANSPORT_FORMAT_PROTOBUF
+
         exporter = ZipkinSpanExporter(
-            service_name=service_name, url=url, ipv4=ipv4, ipv6=ipv6,
+            service_name=service_name,
+            url=url,
+            ipv4=ipv4,
+            ipv6=ipv6,
+            transport_format=transport_format,
         )
 
         self.assertEqual(exporter.service_name, service_name)
@@ -97,9 +121,10 @@ class TestZipkinSpanExporter(unittest.TestCase):
         self.assertEqual(exporter.ipv4, ipv4)
         self.assertEqual(exporter.ipv6, ipv6)
         self.assertEqual(exporter.url, url)
+        self.assertEqual(exporter.transport_format, transport_format)
 
     # pylint: disable=too-many-locals,too-many-statements
-    def test_export(self):
+    def test_export_json(self):
         span_names = ("test1", "test2", "test3", "test4")
         trace_id = 0x6E0C63257DE34C926F9EFCD03927272E
         span_id = 0x34BF92DEEFC58C92
@@ -205,6 +230,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
 
         service_name = "test-service"
         local_endpoint = {"serviceName": service_name, "port": 9411}
+        span_kind = SPAN_KIND_MAP_JSON[SpanKind.INTERNAL]
 
         exporter = ZipkinSpanExporter(service_name)
         expected_spans = [
@@ -215,7 +241,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
                 "timestamp": start_times[0] // 10 ** 3,
                 "duration": durations[0] // 10 ** 3,
                 "localEndpoint": local_endpoint,
-                "kind": None,
+                "kind": span_kind,
                 "tags": {
                     "key_bool": "False",
                     "key_string": "hello_world",
@@ -245,7 +271,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
                 "timestamp": start_times[1] // 10 ** 3,
                 "duration": durations[1] // 10 ** 3,
                 "localEndpoint": local_endpoint,
-                "kind": None,
+                "kind": span_kind,
                 "tags": {
                     "key_resource": "some_resource",
                     "otel.status_code": "1",
@@ -259,7 +285,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
                 "timestamp": start_times[2] // 10 ** 3,
                 "duration": durations[2] // 10 ** 3,
                 "localEndpoint": local_endpoint,
-                "kind": None,
+                "kind": span_kind,
                 "tags": {
                     "key_string": "hello_world",
                     "key_resource": "some_resource",
@@ -274,7 +300,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
                 "timestamp": start_times[3] // 10 ** 3,
                 "duration": durations[3] // 10 ** 3,
                 "localEndpoint": local_endpoint,
-                "kind": None,
+                "kind": span_kind,
                 "tags": {
                     "otel.instrumentation_library.name": "name",
                     "otel.instrumentation_library.version": "version",
@@ -294,6 +320,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
         kwargs = mock_post.call_args[1]
 
         self.assertEqual(kwargs["url"], "http://localhost:9411/api/v2/spans")
+        self.assertEqual(kwargs["headers"]["Content-Type"], "application/json")
         actual_spans = sorted(
             json.loads(kwargs["data"]), key=lambda span: span["timestamp"]
         )
@@ -307,7 +334,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
             self.assertEqual(expected_annotations, actual_annotations)
 
     # pylint: disable=too-many-locals
-    def test_zero_padding(self):
+    def test_export_json_zero_padding(self):
         """test that hex ids starting with 0
         are properly padded to 16 or 32 hex chars
         when exported
@@ -355,7 +382,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
                 "timestamp": start_time // 10 ** 3,
                 "duration": duration // 10 ** 3,
                 "localEndpoint": local_endpoint,
-                "kind": None,
+                "kind": SPAN_KIND_MAP_JSON[SpanKind.INTERNAL],
                 "tags": {"otel.status_code": "1"},
                 "annotations": None,
                 "debug": True,
@@ -383,7 +410,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
         status = exporter.export(spans)
         self.assertEqual(SpanExportResult.FAILURE, status)
 
-    def test_max_tag_length(self):
+    def test_export_json_max_tag_length(self):
         service_name = "test-service"
 
         span_context = trace_api.SpanContext(
@@ -617,3 +644,278 @@ class TestZipkinSpanExporter(unittest.TestCase):
         self.assertEqual(tags["tuple6"], '["True"]')
         self.assertEqual(tags["range1"], '["0", "1"]')
         self.assertEqual(tags["range2"], '["0", "1"]')
+
+    # pylint: disable=too-many-locals,too-many-statements
+    def test_export_protobuf(self):
+        span_names = ("test1", "test2", "test3", "test4")
+        trace_id = 0x6E0C63257DE34C926F9EFCD03927272E
+        span_id = 0x34BF92DEEFC58C92
+        parent_id = 0x1111111111111111
+        other_id = 0x2222222222222222
+
+        base_time = 683647322 * 10 ** 9  # in ns
+        start_times = (
+            base_time,
+            base_time + 150 * 10 ** 6,
+            base_time + 300 * 10 ** 6,
+            base_time + 400 * 10 ** 6,
+        )
+        durations = (50 * 10 ** 6, 100 * 10 ** 6, 200 * 10 ** 6, 300 * 10 ** 6)
+        end_times = (
+            start_times[0] + durations[0],
+            start_times[1] + durations[1],
+            start_times[2] + durations[2],
+            start_times[3] + durations[3],
+        )
+
+        span_context = trace_api.SpanContext(
+            trace_id,
+            span_id,
+            is_remote=False,
+            trace_flags=TraceFlags(TraceFlags.SAMPLED),
+        )
+        parent_span_context = trace_api.SpanContext(
+            trace_id, parent_id, is_remote=False
+        )
+        other_context = trace_api.SpanContext(
+            trace_id, other_id, is_remote=False
+        )
+
+        event_attributes = {
+            "annotation_bool": True,
+            "annotation_string": "annotation_test",
+            "key_float": 0.3,
+        }
+
+        event_timestamp = base_time + 50 * 10 ** 6
+        event = trace.Event(
+            name="event0",
+            timestamp=event_timestamp,
+            attributes=event_attributes,
+        )
+
+        link_attributes = {"key_bool": True}
+
+        link = trace_api.Link(
+            context=other_context, attributes=link_attributes
+        )
+
+        otel_spans = [
+            trace._Span(
+                name=span_names[0],
+                context=span_context,
+                parent=parent_span_context,
+                events=(event,),
+                links=(link,),
+            ),
+            trace._Span(
+                name=span_names[1], context=parent_span_context, parent=None
+            ),
+            trace._Span(
+                name=span_names[2], context=other_context, parent=None
+            ),
+            trace._Span(
+                name=span_names[3], context=other_context, parent=None
+            ),
+        ]
+
+        otel_spans[0].start(start_time=start_times[0])
+        otel_spans[0].resource = Resource({})
+        # added here to preserve order
+        otel_spans[0].set_attribute("key_bool", False)
+        otel_spans[0].set_attribute("key_string", "hello_world")
+        otel_spans[0].set_attribute("key_float", 111.22)
+        otel_spans[0].set_status(
+            Status(StatusCode.ERROR, "Example description")
+        )
+        otel_spans[0].end(end_time=end_times[0])
+
+        otel_spans[1].start(start_time=start_times[1])
+        otel_spans[1].resource = Resource(
+            attributes={"key_resource": "some_resource"}
+        )
+        otel_spans[1].end(end_time=end_times[1])
+
+        otel_spans[2].start(start_time=start_times[2])
+        otel_spans[2].set_attribute("key_string", "hello_world")
+        otel_spans[2].resource = Resource(
+            attributes={"key_resource": "some_resource"}
+        )
+        otel_spans[2].end(end_time=end_times[2])
+
+        otel_spans[3].start(start_time=start_times[3])
+        otel_spans[3].resource = Resource({})
+        otel_spans[3].end(end_time=end_times[3])
+        otel_spans[3].instrumentation_info = InstrumentationInfo(
+            name="name", version="version"
+        )
+
+        service_name = "test-service"
+        local_endpoint = zipkin_pb2.Endpoint(
+            service_name=service_name, port=9411
+        )
+        span_kind = SPAN_KIND_MAP_PROTOBUF[SpanKind.INTERNAL]
+
+        expected_spans = zipkin_pb2.ListOfSpans(
+            spans=[
+                zipkin_pb2.Span(
+                    trace_id=trace_id.to_bytes(
+                        length=16, byteorder="big", signed=False
+                    ),
+                    id=ZipkinSpanExporter.format_pbuf_span_id(span_id),
+                    name=span_names[0],
+                    timestamp=nsec_to_usec_round(start_times[0]),
+                    duration=nsec_to_usec_round(durations[0]),
+                    local_endpoint=local_endpoint,
+                    kind=span_kind,
+                    tags={
+                        "key_bool": "False",
+                        "key_string": "hello_world",
+                        "key_float": "111.22",
+                        "otel.status_code": "2",
+                        "otel.status_description": "Example description",
+                    },
+                    debug=True,
+                    parent_id=ZipkinSpanExporter.format_pbuf_span_id(
+                        parent_id
+                    ),
+                    annotations=[
+                        zipkin_pb2.Annotation(
+                            timestamp=nsec_to_usec_round(event_timestamp),
+                            value=json.dumps(
+                                {
+                                    "event0": {
+                                        "annotation_bool": True,
+                                        "annotation_string": "annotation_test",
+                                        "key_float": 0.3,
+                                    }
+                                }
+                            ),
+                        ),
+                    ],
+                ),
+                zipkin_pb2.Span(
+                    trace_id=trace_id.to_bytes(
+                        length=16, byteorder="big", signed=False
+                    ),
+                    id=ZipkinSpanExporter.format_pbuf_span_id(parent_id),
+                    name=span_names[1],
+                    timestamp=nsec_to_usec_round(start_times[1]),
+                    duration=nsec_to_usec_round(durations[1]),
+                    local_endpoint=local_endpoint,
+                    kind=span_kind,
+                    tags={
+                        "key_resource": "some_resource",
+                        "otel.status_code": "1",
+                    },
+                ),
+                zipkin_pb2.Span(
+                    trace_id=trace_id.to_bytes(
+                        length=16, byteorder="big", signed=False
+                    ),
+                    id=ZipkinSpanExporter.format_pbuf_span_id(other_id),
+                    name=span_names[2],
+                    timestamp=nsec_to_usec_round(start_times[2]),
+                    duration=nsec_to_usec_round(durations[2]),
+                    local_endpoint=local_endpoint,
+                    kind=span_kind,
+                    tags={
+                        "key_string": "hello_world",
+                        "key_resource": "some_resource",
+                        "otel.status_code": "1",
+                    },
+                ),
+                zipkin_pb2.Span(
+                    trace_id=trace_id.to_bytes(
+                        length=16, byteorder="big", signed=False
+                    ),
+                    id=ZipkinSpanExporter.format_pbuf_span_id(other_id),
+                    name=span_names[3],
+                    timestamp=nsec_to_usec_round(start_times[3]),
+                    duration=nsec_to_usec_round(durations[3]),
+                    local_endpoint=local_endpoint,
+                    kind=span_kind,
+                    tags={
+                        "otel.instrumentation_library.name": "name",
+                        "otel.instrumentation_library.version": "version",
+                        "otel.status_code": "1",
+                    },
+                ),
+            ],
+        )
+
+        exporter = ZipkinSpanExporter(
+            service_name, transport_format=TRANSPORT_FORMAT_PROTOBUF
+        )
+        mock_post = MagicMock()
+        with patch("requests.post", mock_post):
+            mock_post.return_value = MockResponse(200)
+            status = exporter.export(otel_spans)
+            self.assertEqual(SpanExportResult.SUCCESS, status)
+
+        # pylint: disable=unsubscriptable-object
+        kwargs = mock_post.call_args[1]
+
+        self.assertEqual(kwargs["url"], "http://localhost:9411/api/v2/spans")
+        self.assertEqual(
+            kwargs["headers"]["Content-Type"], "application/x-protobuf"
+        )
+        self.assertEqual(
+            zipkin_pb2.ListOfSpans.FromString(kwargs["data"]), expected_spans
+        )
+
+    def test_export_protobuf_max_tag_length(self):
+        service_name = "test-service"
+
+        span_context = trace_api.SpanContext(
+            0x0E0C63257DE34C926F9EFCD03927272E,
+            0x04BF92DEEFC58C92,
+            is_remote=False,
+            trace_flags=TraceFlags(TraceFlags.SAMPLED),
+        )
+
+        span = trace._Span(name="test-span", context=span_context,)
+
+        span.start()
+        span.resource = Resource({})
+        # added here to preserve order
+        span.set_attribute("k1", "v" * 500)
+        span.set_attribute("k2", "v" * 50)
+        span.set_status(Status(StatusCode.ERROR, "Example description"))
+        span.end()
+
+        exporter = ZipkinSpanExporter(
+            service_name, transport_format=TRANSPORT_FORMAT_PROTOBUF,
+        )
+        mock_post = MagicMock()
+        with patch("requests.post", mock_post):
+            mock_post.return_value = MockResponse(200)
+            status = exporter.export([span])
+            self.assertEqual(SpanExportResult.SUCCESS, status)
+
+        # pylint: disable=unsubscriptable-object
+        kwargs = mock_post.call_args[1]
+        actual_spans = zipkin_pb2.ListOfSpans.FromString(kwargs["data"])
+        span_tags = actual_spans.spans[0].tags
+
+        self.assertEqual(len(span_tags["k1"]), 128)
+        self.assertEqual(len(span_tags["k2"]), 50)
+
+        exporter = ZipkinSpanExporter(
+            service_name,
+            transport_format=TRANSPORT_FORMAT_PROTOBUF,
+            max_tag_value_length=2,
+        )
+        mock_post = MagicMock()
+        with patch("requests.post", mock_post):
+            mock_post.return_value = MockResponse(200)
+            status = exporter.export([span])
+            self.assertEqual(SpanExportResult.SUCCESS, status)
+
+        # pylint: disable=unsubscriptable-object
+        kwargs = mock_post.call_args[1]
+        actual_spans = zipkin_pb2.ListOfSpans.FromString(kwargs["data"])
+        span_tags = actual_spans.spans[0].tags
+
+        self.assertEqual(len(span_tags["k1"]), 2)
+        self.assertEqual(len(span_tags["k2"]), 2)